re PR bootstrap/53912 (bootstrap fails using default c++ mode in stage 2 and 3 for native x86_64-w64-mingw32)
	PR target/53912
	* ggc-common.c (POINTER_HASH): Cast from pointer via intptr_t.

From-SVN: r194071

+++ gcc/ggc-common.c
@@ -308,7 +308,7 @@ struct ptr_data
   enum gt_types_enum type;
 };
 
-#define POINTER_HASH(x) (hashval_t)((long)x >> 3)
+#define POINTER_HASH(x) (hashval_t)((intptr_t)x >> 3)
 
 /* Register an object in the hash table.  */
